nitorch/test-requirements.txt
numpy
matplotlib
pandas
sklearn
scipy
nibabel
nilearn
h5py
tabulate
tqdm
torchvision
